StatLabIA-backend/Projet-StatLabIA-main/data-cleaning-service/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
pandas==2.2.0
numpy==1.26.0
pymongo
minio==7.1.13
kafka-python==2.1.0
python-dotenv==1.0.0
openpyxl==3.1.3
